HH-218702 move RoutingDataSource to nab-jdbc module
