+++ nab-data-source/src/main/java/ru/hh/nab/datasource/DataSourceSettings.java
@@ -1,8 +1,6 @@
 package ru.hh.nab.datasource;
 
 public final class DataSourceSettings {
-  public static final String DATASOURCE_NAME_FORMAT = "%s.%s";
-
   // validationTimeout = connectionTimeout * DEFAULT_VALIDATION_TIMEOUT_RATIO
   public static final double DEFAULT_VALIDATION_TIMEOUT_RATIO = 0.8;
