+++ nab-jdbc/pom.xml
@@ -23,6 +23,10 @@
             <artifactId>nab-jdbc-common</artifactId>
         </dependency>
         <dependency>
+            <groupId>ru.hh.nab</groupId>
+            <artifactId>nab-metrics</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-jdbc</artifactId>
         </dependency>
